Initial vite commit
