+++ README.md
@@ -1,13 +1,18 @@
-# cells-genetic-algorithm
-Genetic algorithm which simulates 2d cells moving towards a goal (bacteria) and avoiding viruses.
+# prey-predators-genetic-algorithm
 
-Each organism is represented as a circle on the canvas, by default cells are white and the more mutations a cell has the darker
-the cell is. Bacteria is green and the distance of a Cell from the Bacteria or the number of steps it took to reach the
-Bacteria determine the fitness of the Cell. Viruses are blue, if a Cell comes into contact with a virus it dies. The best Cell
-from the previous generation is black. 
+Genetic algorithm which simulates natural selection and genetic diversity in a competitive environment with limited resources.
 
-Each Cell has an energy allowance, for each unit of energy the Cell accelerates in a direction. If the Cell runs out of energy,
-comes into contact with another organism it is considered dead. 
+(Predators are blue and prey are stationary green circles)
 
-It takes approx 200 generations for the population to converge. Different mutation methods and energy allowances for cells 
-can increase or decrease the number of generations requried for convergence. 
+#### Predators Attributes:
+  - A limited amount of energy - this determines how far they can travel before dying. 
+  - Sensory distance - indicated by the black ring, this is the distance that they can detect prey or obstacles in the enviornment.
+  - Velocity - indicated by the black line, direction indicates movement direction and length indicates speed.
+
+#### The simulations works as follows:
+  - If a predator eats one prey and returns to the edge of the map it survies to the next generation.
+  - If a predator eats two prey and returns to the edge of the map it survives and reproduces one offspring with a chance of mutation.
+
+The mutations can be either increased speed, or increased sensory distance. 
+
+Each attribute has an associated cost curve which determines the cost of movement in the world, speed has a higher cost rate than sensory distance so investing in speed is expensive and more risky.
