+++ index.html
@@ -1,15 +1,14 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8"/>
-  <link rel="stylesheet" href="main.css">
-  <script src="https://code.jquery.com/jquery-3.4.1.min.js" integrity="sha256-CSXorXvZcTkaix6Yvo6HppcZGetbYMGWSFlBw8HfCJo=" crossorigin="anonymous"></script>
-</head>
-<body>
-  <h1>Generation: <span id="generation">1</span></h1>
-  <canvas class="cells">Canvas not suppored</canvas>
-  <script type="text/javascript" src="script.js">
-
-  </script>
-</body>
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Vite App</title>
+  </head>
+  <body>
+    <h1>Generation: <span id="generation">1</span></h1>
+    <canvas id="canvas" width="600" height="600"></canvas>
+    <script type="module" src="/src/main.ts"></script>
+  </body>
 </html>
