+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "cells-genetic-algorithm",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^7.8.0",
+    "eslint": "~8.57.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-vitest": "^0.3.22",
+    "prettier": "^3.2.5",
+    "typescript": "~5.4.2",
+    "vite": "~5.0.0",
+    "vite-plugin-dts": "~3.8.1",
+    "vitest": "^1.3.1"
+  }
+}
